Extract initialization functions to utils
Extract the functions to separate files to export them for usage in
ECDSA module.
